Change microcontrast uniformity adjuster range, fixes #5129

+++ rtengine/procparams.h
@@ -558,7 +558,7 @@ struct SharpenMicroParams {
     bool    matrix;
     double  amount;
     double  contrast;
-    double  uniformity;
+    int     uniformity;
 
     SharpenMicroParams();
